refactor: move RosbagManager to lib
